Factor out signal handling into a platform-specific build tag.
