+++ agent/agent.go
@@ -18,13 +18,9 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"runtime"
 	"sync"
 	"time"
 
-	"golang.org/x/sys/unix"
-
-	"github.com/alecthomas/units"
 	cgm "github.com/circonus-labs/circonus-gometrics"
 	"github.com/jackc/pgx"
 	"github.com/joyent/pg_prefaulter/agent/fhcache"
@@ -78,11 +74,7 @@ func New(cfg *config.Config) (a *Agent, err error) {
 	a.metrics.SetTextValue(metricsVersionSelfDate, buildtime.DATE)
 	a.metrics.SetTextValue(metricsVersionSelfVersion, buildtime.VERSION)
 
-	// Handle shutdown via a.shutdownCtx
-	a.signalCh = make(chan os.Signal, 10)
-	signal.Notify(a.signalCh, os.Interrupt, unix.SIGTERM, unix.SIGHUP, unix.SIGPIPE, unix.SIGINFO)
-
-	a.shutdownCtx, a.shutdown = context.WithCancel(context.Background())
+	a.setupSignals()
 
 	if err := a.initDBPool(cfg); err != nil {
 		return nil, errors.Wrap(err, "unable to initialize db connection pool")
@@ -346,35 +338,6 @@ func (a *Agent) getWALFiles() (pg.WALFiles, error) {
 	return walFiles, nil
 }
 
-// handleSignals runs the signal handler thread
-func (a *Agent) handleSignals() {
-	const stacktraceBufSize = 1 * units.MiB
-
-	// pre-allocate a buffer
-	buf := make([]byte, stacktraceBufSize)
-
-	for {
-		select {
-		case <-a.shutdownCtx.Done():
-			log.Debug().Msg("Shutting down")
-			return
-		case sig := <-a.signalCh:
-			log.Info().Str("signal", sig.String()).Msg("Received signal")
-			switch sig {
-			case os.Interrupt, unix.SIGTERM:
-				a.shutdown()
-			case unix.SIGPIPE, unix.SIGHUP:
-				// Noop
-			case unix.SIGINFO:
-				stacklen := runtime.Stack(buf, true)
-				fmt.Printf("=== received SIGINFO ===\n*** goroutine dump...\n%s\n*** end\n", buf[:stacklen])
-			default:
-				panic(fmt.Sprintf("unsupported signal: %v", sig))
-			}
-		}
-	}
-}
-
 // prefaultWALFiles pre-faults the heap pages referenced in WAL files.  When
 // moreWork is set to true it indicates the caller should loop immediately.
 func (a *Agent) prefaultWALFiles(walFiles pg.WALFiles) (moreWork bool, err error) {
@@ -390,7 +353,6 @@ func (a *Agent) prefaultWALFiles(walFiles pg.WALFiles) (moreWork bool, err error
 			waitWALFiles = append(waitWALFiles, walFile)
 		}
 	}
-
 
 	return len(waitWALFiles) > pg.NumOldLSNs, nil
 }
